[Refactor] MainViewController 로직 분리
Contents
- MainViewregistrator 생성
- MainViewDataSourcemanager 생성
