+++ iOS/AirbnbApp/AirbnbApp/ViewController/Main/MainViewController.swift
@@ -28,7 +28,7 @@ final class MainViewController: UIViewController {
     
     private var mockData: [SectionDataSource] = [SectionDataSource.hero(image: HeroImage(image: "https://w7.pngwing.com/pngs/332/451/png-transparent-pepe-the-frog-pepe-4chan-television-face-leaf.png")),
                                                  SectionDataSource.recommend(recommend: Recommend(image: "https://w7.pngwing.com/pngs/332/451/png-transparent-pepe-the-frog-pepe-4chan-television-face-leaf.png", name: "Good House")),
-                                                 SectionDataSource.nearSpot(spot: NearSpot(image: "https://w7.pngwing.com/pngs/332/451/png-transparent-pepe-the-frog-pepe-4chan-television-face-leaf.png", spotName: "좋지 않은 숙소", distance: 30))
+                                                 SectionDataSource.nearSpot(spot: NearSpot(image: "https://w7.pngwing.com/pngs/332/451/png-transparent-pepe-the-frog-pepe-4chan-television-face-leaf.png", spotName: "Come~", distance: 30)),                                                 SectionDataSource.recommend(recommend: Recommend(image: "https://w7.pngwing.com/pngs/332/451/png-transparent-pepe-the-frog-pepe-4chan-television-face-leaf.png", name: "Come on")),                                                 SectionDataSource.nearSpot(spot: NearSpot(image: "https://w7.pngwing.com/pngs/332/451/png-transparent-pepe-the-frog-pepe-4chan-television-face-leaf.png", spotName: "좋지 않은 숙소", distance: 30))
     ]
     
     
@@ -37,7 +37,6 @@ final class MainViewController: UIViewController {
         addViews()
         setLayouts()
         setUpDataSource()
-        snapshot()
         setSearchViewController()
         setNavigationController()
     }
@@ -80,85 +79,18 @@ final class MainViewController: UIViewController {
 extension MainViewController {
     
     func setUpDataSource() {
-        let heroCellRegistration = MainViewRegistrator.createHeroCellRegestration()
-        let nearSpotCellRegistration = MainViewRegistrator.createNearSpotCellRegestration()
-        let recommendResgistration = MainViewRegistrator.createRecommendSpotCellRegestration()
-        let sectionHeaderResigtration = MainViewRegistrator.createHeaderRegistration()
-        
-        let dataSource: UICollectionViewDiffableDataSource<Section, SectionDataSource>? =
-            .init(collectionView: collectionView) { collectionView, indexPath, data in
-                guard let section = Section(rawValue: indexPath.section) else { return  nil }
-                
-                switch section {
-                case .hero:
-                    return collectionView.dequeueConfiguredReusableCell(
-                        using: heroCellRegistration,
-                        for: indexPath,
-                        item: HeroImage(image: data.image))
-                case .nearSpot:
-                    return collectionView.dequeueConfiguredReusableCell(
-                        using: nearSpotCellRegistration,
-                        for: indexPath,
-                        item: NearSpot(image: data.image,
-                                       spotName: data.spotName ?? "",
-                                       distance: data.distance ?? 0))
-                case .recommend:
-                    return collectionView.dequeueConfiguredReusableCell(
-                        using: recommendResgistration,
-                        for: indexPath,
-                        item: Recommend(image: data.image,
-                                        name: data.spotName ?? "")
-                )
-            }
-        }
-        
-        dataSource?.supplementaryViewProvider = { collectionView, _, indexPath in
-            collectionView.dequeueConfiguredReusableSupplementary(using: sectionHeaderResigtration, for: indexPath)
-        }
- 
-        self.dataSource = dataSource
-    }
-    
-    private func snapshot() {
-        
-        var heroImages = [SectionDataSource]()
-        var nearSpots = [SectionDataSource]()
-        var recommends = [SectionDataSource]()
-        
-        mockData.forEach { (data: SectionDataSource) in
-            switch data {
-            case .hero:
-                heroImages.append(data)
-            case .nearSpot:
-                nearSpots.append(data)
-            case .recommend:
-                recommends.append(data)
-            }
-        }
-        
-        var heroImageSnapShot = NSDiffableDataSourceSectionSnapshot<SectionDataSource>()
-        heroImageSnapShot.append(heroImages)
-            
-        var nearSpotSnapShot = NSDiffableDataSourceSectionSnapshot<SectionDataSource>()
-        nearSpotSnapShot.append(nearSpots)
-        
-        var recommendSnapShot = NSDiffableDataSourceSectionSnapshot<SectionDataSource>()
-        recommendSnapShot.append(recommends)
-        
-        self.dataSource?.apply(heroImageSnapShot, to: .hero, animatingDifferences: true)
-        self.dataSource?.apply(nearSpotSnapShot, to: .nearSpot, animatingDifferences: true)
-        self.dataSource?.apply(recommendSnapShot, to: .recommend, animatingDifferences: true)
-        
+        MainViewDataSourceManager.setDataSource(in: collectionView)
+        MainViewDataSourceManager.snapshot(data: mockData)
     }
 }
 
-private enum Section: Int, CaseIterable {
+enum Section: Int, CaseIterable {
     case hero
     case nearSpot
     case recommend
 }
 
-private enum SectionDataSource: Hashable {
+enum SectionDataSource: Hashable {
     case hero(image: HeroImage)
     case nearSpot(spot: NearSpot)
     case recommend(recommend: Recommend)
