+++ iOS/AirbnbApp/AirbnbApp/ViewController/Main/SectionDataModel/HeroImage.swift
