+++ iOS/AirbnbApp/AirbnbApp/ViewController/Main/ViewModel/SectionHeaderViewModel.swift
@@ -10,7 +10,7 @@ import Foundation
 final class SectionHeaderViewModel {
     private let heraderTitle:[String?] = [nil,
                                           "가까운 여행지 둘러보기",
-                                          "어디에서나, 여행은 살아보는거야!"]
+                                          "어디에서나,여행은\n살아보는거야!"]
     
     
     func getTitle(at sectionNumber:Int) -> String? {
